Extend syntax for common log fields
