+++ log_middleware_test.go
@@ -2,10 +2,11 @@ package logging
 
 import (
 	"bytes"
-	"github.com/stretchr/testify/assert"
 	"net/http"
 	"net/http/httptest"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func Test_LogMiddleware_Panic(t *testing.T) {
@@ -13,7 +14,7 @@ func Test_LogMiddleware_Panic(t *testing.T) {
 
 	// given: a logger
 	b := bytes.NewBuffer(nil)
-	Logger.Out = b
+	Log.Out = b
 
 	// and a handler which raises a panic
 	lm := NewLogMiddleware(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -37,7 +38,7 @@ func Test_LogMiddleware_Log_implicit200(t *testing.T) {
 
 	// given: a logger
 	b := bytes.NewBuffer(nil)
-	Logger.Out = b
+	Log.Out = b
 
 	// and a handler which gets an 200er code implicitly
 	lm := NewLogMiddleware(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -60,7 +61,7 @@ func Test_LogMiddleware_Log_404(t *testing.T) {
 
 	// given: a logger
 	b := bytes.NewBuffer(nil)
-	Logger.Out = b
+	Log.Out = b
 
 	// and a handler which gets an 404er code explicitly
 	lm := NewLogMiddleware(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
